Move funtions to new files
